+++ pgrx/src/lib.rs
@@ -69,7 +69,6 @@ pub mod nullable;
 pub mod palloc;
 pub mod pg_catalog;
 pub mod pgbox;
-pub mod pglz;
 pub mod rel;
 pub mod shmem;
 pub mod spi;
